Extract table initialization to separate component
